refactor(dev): devServer_unstable namespace
